// scripts/importFromOldDB/libs/API.ts
export type CreateSheetInput = {
  id?: string | null,
  sub: string,
  year: number,
  companyID: string,
  groupID: string,
  grade: string,
  careerPlan?: string | null,
  careerPlanComment?: string | null,
  reviewComment?: string | null,
  reviewDate?: string | null,
  selfCheckDate?: string | null,
  firstComment?: string | null,
  firstCheckDate?: string | null,
  secondComment?: string | null,
  secondCheckDate?: string | null,
  overAllEvaluation?: number | null,
  statusValue?: number | null,
  interviewPlanDate?: string | null,
  interviewPlanComment?: string | null,
  InterviewMid1Date?: string | null,
  InterviewMid1Comment?: string | null,
  InterviewMid2Date?: string | null,
  InterviewMid2Comment?: string | null,
  InterviewMid3Date?: string | null,
  InterviewMid3Comment?: string | null,
  revieweeUsername: string,
  secondUsername: string,
  sheetGroupName?: string | null,
  referencer?: Array< string > | null,
  reviewee: string,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
};